Work on understanding stuff
Why? Because when I add a new book to the myLibrary array,
the eventListener on the "Remove" button does not do anything
anymore. Before that it displays the id of the button clicked.
We need to know which button was clicked in order to remove the
right book in the future.